Collect girafe trees
We've started refactoring the way we perform our automatic tree patching. The idea here is that we build up an entire tree database first - which is committed to version control - and then we can use this tree database to perform the patching. This decouples the discovery of the models from the patching itself, which should make the code more readable.