Refactor database structure
The tree database now stores *all* models it finds in each and every exemplar. That way we can easily look up what tree model is found in which exemplar instead of having to regenerate the id again. This means that we can now change ids by hand, and not have to update the id generating function.